stability: implement node liveness; first step towards new range leases
This change adds a node liveness table as a global system table.
Nodes periodically write updates to their liveness record by doing
a conditional put to the liveness table. The leader of the range
containing the node liveness table gossips the latest information
to the rest of the system.

Each node has a `NodeLiveness` object which can be used to query
the status of any other node to find out if it's live or non-live
according to the liveness threshold duration compared to the last
time it successfully heartbeat its liveness record.

The as-yet-unused `IncrementEpoch` mechanism is also added in this
PR, for eventual use with the planned epoch-based range leader leases.

Updated the range leader lease RFC to reflect current thinking.

+++ server/status.go
@@ -509,7 +509,7 @@ func (s *statusServer) Node(ctx context.Context, req *serverpb.NodeRequest) (*st
 		return nil, grpc.Errorf(codes.InvalidArgument, err.Error())
 	}
 
-	key := keys.NodeStatusKey(int32(nodeID))
+	key := keys.NodeStatusKey(nodeID)
 	b := &client.Batch{}
 	b.Get(key)
 	if err := s.db.Run(s.ctx, b); err != nil {
